// earnaha/src/exceptions/error-code.enum.ts
// exceptions/error-code.enum.ts
export enum ErrorCode {
  SUCCESS = 1000,
  USER_ALREADY_EXISTS = 1001,
  INVALID_CREDENTIALS = 1002,
  USER_NOT_FOUND = 1003,
  ACCOUNT_CONFLICT_GOOGLE = 1004,
  ACCOUNT_CONFLICT_EMAIL = 1005,
  INVALID_MERGE_KEY = 1006,
  NO_VALID_SESSION = 1007,
  INVALID_PASSWORD = 1008,
  VERIFY_LIMIT = 1009,
  REDUNDANT_CHANGE_PASS = 1010,
  UNKNOWN_ERROR = 9999,
}